+++ expressServer/index.js
@@ -57,6 +57,7 @@ app.get('/api/users', async (req, res) => {
     await connectDB();
     const users = await User.find({ username: { $regex: search, $options: 'i' } });
     res.status(200).json(users);
+
   } catch (error) {
     console.error('Error fetching users:', error);
     res.status(500).json({ message: 'Internal server error' });
