--- expressServer/routes/loginFormAPI.js
@@ -1,25 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-router.post('/login', async (req, res) => {
-
-  const { username, password } = req.body;
-
-  try {
-    const user = await User.findOne({ username });
-
-    if (!user) {
-      return res.status(404).json({ message: 'User not found' });
-    }
-    if (user.password !== password) {
-      console.log('Error during login:');
-      return res.status(401).json({ message: 'Incorrect password' });
-    }
-    res.status(200).json({ message: 'Login successful' });
-  } catch (error) {
-    console.error('Error during login:', error);
-    res.status(500).json({ message: 'Internal server error' });
-  }
-});
-
-module.exports = router;
