+++ reactLogin/src/App.tsx
@@ -1,31 +1,12 @@
-import React, { useState } from 'react';
-import axios from 'axios';
+import React from 'react';
 import './App.css';
 import LoginForm from './components/LoginForm';
-import { useNavigate } from 'react-router-dom';
 
 const App: React.FC = () => {
-  const [message, setMessage] = useState<string>('');
-  const navigate = useNavigate();
-
-  const handleLoginSubmit = async (values: { username: string; password: string }) => {
-    try {
-      const response = await axios.post('/login', values);
-      setMessage(response.data.message);
-      navigate('/');
-    } catch (error) {
-      if (axios.isAxiosError(error)) {
-        setMessage(error.response?.data.message || 'Login failed. Please try again.');
-      } else {
-        setMessage('An unknown error occurred.');
-      }
-    }
-  };
 
   return (
     <div className="App">
-      <LoginForm onSubmit={handleLoginSubmit} />
-      {message && <p>{message}</p>}
+      <LoginForm />
     </div>
   );
 };
